front: move requests display in a component (#776)
